pyproject.toml
[tool.poetry]
name = "mocli"
version = "0.1.0"
description = ""
authors = ["Thomas Basche <thomas.basche@wolt.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.3.3"
attrs = "^23.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
